pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nichi"
version = "2.0.0"
description = "A comprehensive tool for organizing video files, converting VTT subtitles to SRT format, and translating SRT files using Google Gemini AI."
readme = "README.md"
authors = [
    {name = "Rafli", email = "hidayattul.rafli@gmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Utilities",
]
keywords = ["video", "subtitles", "translation", "organization", "vtt", "srt", "gemini"]
requires-python = ">=3.8"
dependencies = [
    "markdown-it-py==4.0.0",
    "mdurl==0.1.2",
    "Pygments==2.19.2",
    "rich==14.1.0",
    "google-generativeai==0.3.0",
    "python-dotenv==1.0.0",
]

[project.urls]
Homepage = "https://github.com/hdytrfli/nichi"
Repository = "https://github.com/hdytrfli/nichi"
Issues = "https://github.com/hdytrfli/nichi/issues"

[project.scripts]
nichi = "nichi.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 120
target-version = ['py38']

[tool.flake8]
max-line-length = 120